Factor out email_react_rendered_content_with_attachment.
